refactor(Progress): rename to ts
