+++ components/progress/mobile/index.tsx
