+++ components/progress/style.ts
