+++ components/progress/view/progress-circle.tsx
@@ -35,8 +35,10 @@ export default class Circle extends Component {
         if (this.underlay && this.overlay) {
             // eslint-disable-next-line
             this.setState({
-                underlayStrokeWidth: this._getCssValue(this.underlay, 'stroke-width') || DEFAULT_STROKE_WIDTH,
-                overlayStrokeWidth: this._getCssValue(this.overlay, 'stroke-width') || DEFAULT_STROKE_WIDTH,
+                underlayStrokeWidth:
+                    this._getCssValue(this.underlay, 'stroke-width') || DEFAULT_STROKE_WIDTH,
+                overlayStrokeWidth:
+                    this._getCssValue(this.overlay, 'stroke-width') || DEFAULT_STROKE_WIDTH,
             });
         }
     }
@@ -60,14 +62,17 @@ export default class Circle extends Component {
     _computeOverlayStrokeDashOffset() {
         const { underlayStrokeWidth, overlayStrokeWidth } = this.state;
         const overlayRadius =
-            HALF_VIEWBOX_WIDTH - overlayStrokeWidth / 2 - (underlayStrokeWidth - overlayStrokeWidth) / 2;
+            HALF_VIEWBOX_WIDTH -
+            overlayStrokeWidth / 2 -
+            (underlayStrokeWidth - overlayStrokeWidth) / 2;
         const overlayLen = Math.PI * 2 * overlayRadius;
         return ((VIEWBOX_WIDTH - this.props.percent) / VIEWBOX_WIDTH) * overlayLen;
     }
 
     _getPath(radius) {
-        return `M ${HALF_VIEWBOX_WIDTH},${HALF_VIEWBOX_WIDTH} m 0,-${radius} a ${radius},${radius} 0 1 1 0,${2 *
-            radius} a ${radius},${radius} 0 1 1 0,-${2 * radius}`;
+        return `M ${HALF_VIEWBOX_WIDTH},${HALF_VIEWBOX_WIDTH} m 0,-${radius} a ${radius},${radius} 0 1 1 0,${
+            2 * radius
+        } a ${radius},${radius} 0 1 1 0,-${2 * radius}`;
     }
 
     render() {
@@ -92,7 +97,9 @@ export default class Circle extends Component {
 
         // overlay path (为居中，减去相对于underlay的宽度)
         const overlayRadius =
-            HALF_VIEWBOX_WIDTH - overlayStrokeWidth / 2 - (underlayStrokeWidth - overlayStrokeWidth) / 2;
+            HALF_VIEWBOX_WIDTH -
+            overlayStrokeWidth / 2 -
+            (underlayStrokeWidth - overlayStrokeWidth) / 2;
         const overlayPath = this._getPath(overlayRadius);
         const overlayLen = Math.PI * 2 * overlayRadius;
         const overlayStrokeDasharray = `${overlayLen}px ${overlayLen}px`;
@@ -111,7 +118,8 @@ export default class Circle extends Component {
             [`${prefix}progress-circle-overlay`]: true,
             [`${prefix}progress-circle-overlay-${state}`]: !color && !progressive && state,
             [`${prefix}progress-circle-overlay-started`]: !color && progressive && percent <= 30,
-            [`${prefix}progress-circle-overlay-middle`]: !color && progressive && percent > 30 && percent < 80,
+            [`${prefix}progress-circle-overlay-middle`]:
+                !color && progressive && percent > 30 && percent < 80,
             [`${prefix}progress-circle-overlay-finishing`]: !color && progressive && percent >= 80,
         });
 
@@ -145,7 +153,9 @@ export default class Circle extends Component {
                         stroke={color}
                     />
                 </svg>
-                {suffixText ? <div className={`${prefix}progress-circle-text`}>{suffixText}</div> : null}
+                {suffixText ? (
+                    <div className={`${prefix}progress-circle-text`}>{suffixText}</div>
+                ) : null}
             </div>
         );
     }
